feat: add delta sse & gunicorn
