+++ Dockerfile
@@ -13,15 +13,17 @@ RUN mkdir hf && uv venv
 
 # Populate pip and uv caches and resolve wheels (no project install) using BuildKit cache mounts
 # (requires BuildKit/Buildx in CI, which the workflow config already sets up)
-RUN --mount=type=cache,target=/root/.cache/pip \ 
+RUN --mount=type=cache,target=/root/.cache/pip \
 	--mount=type=cache,target=/root/.cache/uv \
 	uv sync --link-mode=copy --frozen --no-install-project --no-dev
 
 ADD . .
 
 # Install the project (uses cached wheels from previous step) and verify lockfile
-RUN --mount=type=cache,target=/root/.cache/pip \ 
+RUN --mount=type=cache,target=/root/.cache/pip \
 	--mount=type=cache,target=/root/.cache/uv \
 	uv sync --link-mode=copy --no-dev && uv lock --check
 
+
+# replace for api server: ["uvicorn", "api_server:app", "--workers", "10", "--loop", "uvloop"]
 CMD ["uv", "run", "inky-mbta-tracker"]
