+++ pyproject.toml
@@ -62,6 +62,9 @@ dependencies = [
     "huggingface-hub>=0.34.4",
     "jax>=0.5.3",
     "tqdm>=4.67.1",
+    "json-delta>=2.0.2",
+    "uvicorn-worker>=0.4.0",
+    "gunicorn>=23.0.0",
 ]
 
 [dependency-groups]
@@ -74,6 +77,7 @@ dev = [
 
 [project.scripts]
 inky-mbta-tracker = "main:run_main"
+api-server = "api_server:run_main"
 
 [tool.ruff]
 # Exclude a variety of commonly ignored directories.
